# collection.openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "postman_collection",
    "version": "1.0.0",
    "contact": {}
  },
  "servers": [
    {
      "url": "http://localhost"
    }
  ],
  "paths": {
...
    "/api/hotels/{hotel_id}/room-types/{room_type_id}": {
      "delete": {
        "summary": "Delete Room-Type",
        "description": "### Delete Hotel Room Types\n\n**Genereated using PostbotAI\n\nThis endpoint sends an HTTP DELETE request to remove the room types associated with a specific hotel.\n\n#### Response\n\nThe response for this request is a JSON schema representing the structure of the data returned upon successful deletion of the room types.",
        "operationId": "deleteRoomType",
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "patch": {
        "summary": "Update Room-Type",
        "description": "# Update Room Type Details\n\nThis endpoint is used to update the details of a specific room type for a hotel.\n\n**Genereated using PostbotAI\n\n## Request\n\n- Method: PATCH\n    \n- URL: `http://localhost:3000/api/hotels/3/room-types/1`\n    \n- Headers:\n    \n    - Content-Type: application/json\n        \n- { \"pricePerNight\": 220, \"availableRooms\": 3}\n    \n\n## Response\n\nThe response for this request follows the JSON schema below:\n\n``` json\n{\n  \"type\": \"object\",\n  \"properties\": {\n    \"success\": {\n      \"type\": \"boolean\"\n    },\n    \"message\": {\n      \"type\": \"string\"\n    },\n    \"data\": {\n      \"type\": \"object\",\n      \"properties\": {\n        \"id\": {\n          \"type\": \"integer\"\n        },\n        \"pricePerNight\": {\n          \"type\": \"number\"\n        },\n        \"availableRooms\": {\n          \"type\": \"integer\"\n        }\n        // Additional properties based on the response data\n      }\n    }\n  }\n}\n\n ```",
        "operationId": "updateRoomType",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "availableRooms": {
                    "type": "number",
                    "example": 3
                  },
                  "pricePerNight": {
                    "type": "number",
                    "example": 220
                  }
                }
              },
              "examples": {
                "Update Room-Type": {
                  "value": {
                    "availableRooms": 3,
                    "pricePerNight": 220
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        }
      },
      "parameters": [
        {
          "name": "hotel_id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        },
        {
          "name": "room_type_id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ]
    },
...
    "/api/hotels/{hotel_id}/availability": {
      "get": {
        "summary": "Get-Room-Availability",
        "description": "This endpoint retrieves the availability of a specific hotel for a given date range.\n\n**Genereated using PostbotAI\n\n### Request\n\n#### Request Parameters\n\n- `HOTEL_ID` (path parameter): The unique identifier of the hotel.\n    \n- `startDate` (query parameter): The start date of the availability search (format: YYYY-MM-DD).\n    \n- `endDate` (query parameter): The end date of the availability search (format: YYYY-MM-DD).\n    \n\n### Response\n\nThe response will include the availability status of the hotel for the specified date range.",
        "operationId": "getRoomAvailability",
        "parameters": [
          {
            "name": "startDate",
            "in": "query",
            "schema": {
              "type": "string",
              "example": "2025-04-01"
            }
          },
          {
            "name": "endDate",
            "in": "query",
            "schema": {
              "type": "string",
              "example": "2025-04-05"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      },
      "parameters": [
        {
          "name": "hotel_id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ]
    },
...
    "/api/bookings": {
      "post": {
        "summary": "Create Booking Itinerary",
        "description": "Create Booking Itinerary",
        "operationId": "createBookingItinerary",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "flightBooking": {
                    "type": "string",
                    "example": "AE6838"
                  },
                  "hotelBooking": {
                    "type": "number",
                    "example": 2
                  }
                }
              },
              "examples": {
                "Create Booking Itinerary": {
                  "value": {
                    "flightBooking": "AE6838",
                    "hotelBooking": 2
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/bookings/cancel": {
      "post": {
        "summary": "Cancel booking itinerary",
        "description": "### Cancel Booking\n\n**Genereated using PostbotAI\n\nThis endpoint is used to cancel a hotel booking.\n\n#### Request Body\n\n- `hotelBooking`: The ID of the hotel booking to be canceled.\n    \n\n#### Response\n\nUpon successful cancellation, the server returns a 200 status with a JSON response containing a message indicating the status of the cancellation.",
        "operationId": "cancelBookingItinerary",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "hotelBooking": {
                    "type": "number",
                    "example": 1
                  }
                }
              },
              "examples": {
                "Cancel booking itinerary": {
                  "value": {
                    "hotelBooking": 1
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/bookings/pdf": {
      "get": {
        "summary": "Generate PDF",
        "description": "**Generated using Postbot AI\n\nThis endpoint is used to retrieve a PDF file for a specific itinerary by providing the itinerary ID as a query parameter in the URL.\n\n### Request\n\n- Method: GET\n    \n- Endpoint: [http://localhost:3000/api/bookings/pdf](http://localhost:3000/api/bookings/pdf)\n    \n- Query Parameters:\n    \n    - itineraryId (number, required): The ID of the itinerary for which the PDF file is to be retrieved.\n        \n\n### Response\n\n- Status: 200\n    \n- Content-Type: application/json\n    \n- Body:\n    \n    - url (string): The URL of the PDF file.\n        \n    - message (string): A message related to the retrieval process.\n        \n\nThe response returned is a JSON object with the following schema:\n\n``` json\n{\n    \"type\": \"object\",\n    \"properties\": {\n        \"url\": {\n            \"type\": \"string\"\n        },\n        \"message\": {\n            \"type\": \"string\"\n        }\n    }\n}\n\n ```\n\nThe response includes the URL for the retrieved PDF and a message field for any additional information.",
        "operationId": "generatePdf",
        "parameters": [
          {
            "name": "itineraryId",
            "in": "query",
            "schema": {
              "type": "string",
              "example": "1"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}